Address review feedback: move loader to a helper module with collision-safe names
