+++ tests/test_generate_latest_json.py
@@ -30,7 +30,7 @@ from tempfile import TemporaryDirectory
 from typing import Any
 
 import pytest
-from conftest import load_script_module
+from script_loader import load_script_module
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
 SCRIPT_PATH = REPO_ROOT / ".github" / "scripts" / "generate_latest_json.py"
